Data Engineering - eCommerce/data-simulator/app/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.28